Update patch set 1
Patch Set 1:

> What do you think about renaming all sprint posts like
 > sprint-YYYYMMDD ?

The current naming convention is purely proper to RHOS and makes little sense in the outside world. We can get back to the previous convention (YY-WW where WW is the sprint's end's week number) but your convention makes as much sense, as long as we agree on whether to use the starting date or the ending date.

Patch-set: 1