Update patch set 10
Patch Set 10:

(15 comments)

I expect this post to follow the same spirit as the others in the series. I mean a reader should be able to play (easily) the scenarios on a SF sandbox. Please see my comments, mainly your scenario lack a test, a job, and example where there is a test failure and a test success.

Patch-set: 10
Label: Code-Review=0